% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/precipitationData.R
\docType{data}
\name{DWD_Potsdam_HourlyPrecipitation}
\alias{DWD_Potsdam_HourlyPrecipitation}
\title{DWD historical precipitation data from Potsdam (STATIONS_ID 03987)}
\format{A data frame with 7 columns.
STATIONS_ID (integer: 3987), DWD station ID;
MESS_DATUM (integer: YYYYMMDDHH), timestamp format of measurement;
QN_8 (integer), quality flag;
R1 (numeric), hourly precipitation height in mm;
RS_IND (integer), binary index, 0 no precipitation, 1 precipitation;
WRTR (integer), WR-code for form of precipitation
eor (character), end of record, terminating column;
A detailed explanation can be found under \href{ftp://ftp-cdc.dwd.de/pub/CDC/observations_germany/climate/hourly/precipitation/historical/DESCRIPTION_obsgermany_climate_hourly_precipitation_historical_en.pdf}{Data set desription}}
\source{
Data can be downloaded from \href{ftp://ftp-cdc.dwd.de/pub/CDC/observations_germany/climate/hourly/precipitation/historical/}{DWD Open Data Server}.
Terms and conditions can be found under \href{https://www.dwd.de/EN/service/copyright/copyright_artikel.html}{Copyright}.
}
\usage{
data(DWD_Potsdam_HourlyPrecipitation)
}
\description{
Historical precipitation data from DWD station in Potsdam, Germany.
Values have hourly resolution.
}
\examples{
temperatureSpectrum <- calculateSpectrum(DWD_Potsdam_HourlyPrecipitation$R1[1:2^15], timeStep = 1 / 24)
}
\references{

}
\keyword{datasets}
\keyword{precipitation}
